Move LinkUpdater to updater package
